Remove view and domain. Move tree into gallery
