+++ src/dragAndDrop/index.ts
@@ -2,20 +2,32 @@
 import { dom, div, style } from "../browser";
 import { spacings } from "../designSystem";
 import { colors } from "../designSystem";
+import { AppEvents } from "../events";
 import { itemsStore } from "../globals";
-import { viewIconFor } from "../tree";
+import { Item } from "../items";
+import { viewIconFor } from "../gallery/tree";
 // import FolderIcon from "./FolderIcon";
 
 export class Dnd {
-  itemBeingDragged?: MyItem;
-  itemViewUnder?: MyItem;
+  itemBeingDragged?: Item;
+  itemViewUnder?: Item;
   initialMousePosition: Vector = { x: 0, y: 0 };
   dragAvatar?: HTMLElement;
   dragDestination?: HTMLElement;
   dropPlacement?: DropPlacement;
   appendItemTo?: Element;
 
-  onItemMouseDown = (itemModel: MyItem, e: MouseEvent) => {
+  constructor(events: AppEvents) {
+    events.on("item.mouseDownOnIcon", ({ item, event }) =>
+      this.onItemMouseDown(item, event)
+    );
+
+    events.on("item.mouseMoveOverItem", ({ item, event }) =>
+      this.onItemMouseMoveOver(item, event)
+    );
+  }
+
+  onItemMouseDown = (itemModel: Item, e: MouseEvent) => {
     this.appendItemTo = dom.getFirstElementWithClass(document.body, "app");
 
     this.initialMousePosition = getScreenPosition(e);
@@ -33,7 +45,7 @@ export class Dnd {
       if (!dragAvatar) {
         const dist = distance(initialMousePosition, getScreenPosition(e));
         if (dist > 5) {
-          const icon = div({}); // viewIconFor(itemBeingDragged);
+          const icon = viewIconFor(itemBeingDragged);
           this.dragAvatar = div({ className: "item-dragAvatar" }, [icon]);
           dom.addClass(icon, "item-icon-video");
           this.appendItemTo?.appendChild(this.dragAvatar);
@@ -47,14 +59,14 @@ export class Dnd {
     }
   };
 
-  onItemMouseMoveOver = (itemViewUnder: MyItem, e: MouseEvent) => {
+  onItemMouseMoveOver = (itemViewUnder: Item, e: MouseEvent) => {
     if (this.dragAvatar) {
       this.itemViewUnder = itemViewUnder;
       if (!this.dragDestination) {
         const bulp = dom.div({ className: "item-dragDestinationBulp" });
-        const type = itemViewUnder.type;
         bulp.style.borderRadius =
-          (type === "folder" || type === "YTchannel" ? 5 : 1) + "px";
+          (itemViewUnder.isFolder() || itemViewUnder.isPlaylist() ? 5 : 1) +
+          "px";
 
         this.dragDestination = div({ className: "item-dragDestinationLine" }, [
           bulp,
@@ -118,8 +130,7 @@ export class Dnd {
     const { itemBeingDragged, itemViewUnder, dropPlacement } = this;
     if (itemBeingDragged && itemViewUnder && dropPlacement) {
       if (itemBeingDragged != itemViewUnder)
-        itemsStore.moveItem({
-          itemOver: itemBeingDragged,
+        itemBeingDragged.moveItem({
           placement: dropPlacement,
           itemUnder: itemViewUnder,
         });
