+++ src/lodash/index.ts
@@ -1,0 +1,1 @@
+export * as array from "./array";
